+++ R/rstan_package_skeleton.R
@@ -20,32 +20,42 @@
 #'
 #' This function is very similar to \code{\link[utils]{package.skeleton}} but is
 #' designed for source packages that want to include Stan Programs that can be
-#' built into binary versions.
+#' built into binary versions. See \strong{Details} for a few ways that it
+#' differs from \code{package.skeleton}.
 #'
 #' @export
-#' @param name,list,environment,path,force,code_files Same as in
+#' @param name,list,environment,path,force,code_files See
 #'   \code{\link[utils]{package.skeleton}}.
 #' @param stan_files A character vector with paths to \code{.stan} files to
-#'   include in the package. Otherwise similar to \code{code_files}.
+#'   include in the package (these files will be included in the
+#'   \code{src/stan_files} directory). Otherwise similar to the
+#'   \code{code_files} argument.
 #' @param travis Should a \code{.travis.yml} file be added to the package
 #'   directory? Defaults to \code{TRUE}.
 #'
 #' @details This function first calls \code{\link[utils]{package.skeleton}} and
-#'   then adds the files listed in \code{stan_files} to an exec directory.
-#'   Finally, it downloads several files from the \pkg{rstanarm} package's
+#'   then adds the files listed in \code{stan_files} to the
+#'   \code{src/stan_files} directory. Finally, it downloads several files from
+#'   the \pkg{rstanarm} package's
 #'   \href{http://github.com/stan-dev/rstanarm}{GitHub repository} to facilitate
 #'   building the resulting package. Note that \pkg{\link[rstanarm]{rstanarm}}
 #'   is licensed under the GPL >= 3, so package builders who do not want to be
 #'   governed by that license should not use the downloaded files that contain
-#'   \R code. Otherwise, it may be worth considering whether it would be easier
-#'   to include your \code{.stan} programs and supporting \R code in the
-#'   \pkg{rstanarm} package.
+#'   \R code (that said, \pkg{Rcpp} is GPL, so not using the \pkg{rstanarm}
+#'   files is not the only thing impeding use of other licenses). Otherwise, it
+#'   may be worth considering whether it would be easier to include your
+#'   \code{.stan} programs and supporting \R code in the \pkg{rstanarm} package.
+#'
+#'   Unlike \code{package.skeleton}, \code{rstan_package_skeleton} also creates
+#'   a file in the \code{R/} directory called "\code{name}-package.R", where
+#'   \code{name} is the package name. In this file \code{rstan_package_skeleton}
+#'   writes lines (using \pkg{roxygen2} tags) for ensuring that some necessary
+#'   content makes it into the \code{NAMESPACE} file. Before terminating,
+#'   \code{rstan_package_skeleton} will run \code{roxygen2::roxygenise} so that
+#'   the NAMESPACE is created.
 #'
 #'   After running \code{rstan_package_skeleton} see the
-#'   \code{Read-and-delete-me} file created in the package directory. The
-#'   content in that file contains the content of the \code{Read-and-delete-me}
-#'   file created by \code{\link[utils]{package.skeleton}} plus additional
-#'   Stan-specific instructions.
+#'   \code{Read-and-delete-me} file created in the package directory.
 #'
 #' @seealso
 #' \itemize{
@@ -55,7 +65,7 @@
 #' @template seealso-get-help
 #' @template seealso-useR2016-video
 #'
-#' @importFrom utils download.file packageVersion
+#' @importFrom utils download.file packageVersion available.packages
 #'
 rstan_package_skeleton <-
   function(name = "anRpackage",
@@ -67,6 +77,12 @@ rstan_package_skeleton <-
            stan_files = character(),
            travis = TRUE) {
 
+    if (!requireNamespace("roxygen2", quietly = TRUE)) {
+      stop("Please install the 'roxygen2' package to use this function.")
+    }
+
+    message("Creating package skeleton for package: ", name, domain = NA)
+
     if (length(stan_files) > 0 && !all(grepl("\\.stan$", stan_files)))
       stop("All files named in 'stan_files' must end ",
            "with a '.stan' extension.")
@@ -74,6 +90,14 @@ rstan_package_skeleton <-
     mc <- match.call()
     mc$stan_files <- NULL
     mc[[1]] <- quote(utils::package.skeleton)
+
+    if (is.null(mc$environment)) {
+      has_objects <- length(ls(envir = environment))
+      if (!has_objects) {
+        mc$environment <- new.env(parent = emptyenv())
+        mc$environment$delete_data <- "placeholder data to avoid package.skeleton error"
+      }
+    }
 
     message("Running package.skeleton ...", domain = NA)
     suppressMessages(eval(mc))
@@ -132,21 +156,12 @@ rstan_package_skeleton <-
       destfile = file.path(SRC, "Makevars"),
       quiet = TRUE
     )
-    system2("sed", args = paste0("-i.bak 's@^SOURCES.*@", "SOURCES = ",
-                                 paste("stan_files", stan_files, sep = "/", collapse = " "),
-                                 "@g' ", file.path(SRC, "Makevars")),
-            stdout = FALSE, stderr = FALSE)
-    file.remove(file.path(SRC, "Makevars.bak"))
     download.file(
       .rstanarm_path("src/Makevars.win"),
       destfile = file.path(SRC, "Makevars.win"),
       quiet = TRUE
     )
-    system2("sed", args = paste0("-i.bak 's@^SOURCES.*@", "SOURCES = ",
-                                 paste("stan_files", stan_files, sep = "/", collapse = " "),
-                                 "@g' ", file.path(SRC, "Makevars.win")),
-            stdout = FALSE, stderr = FALSE)
-    file.remove(file.path(SRC, "Makevars.win.bak"))
+
     # register cpp (src/init.cpp)
     init_cpp(name, path = DIR)
 
@@ -176,6 +191,13 @@ rstan_package_skeleton <-
     cat("// Insert all #include<foo.hpp> statements here",
         file = file.path(INCLUDE, "meta_header.hpp"), sep = "\n")
 
+
+    message("Cleaning up unused files ...", domain = NA)
+    .remove_unused_files(DIR)
+
+    message("Updating DESCRIPTION ...", domain = NA)
+    .update_description_file(DIR)
+
     message("Updating R directory ...", domain = NA)
     R <- file.path(DIR, "R")
     dir.create(R, showWarnings = FALSE)
@@ -197,13 +219,16 @@ rstan_package_skeleton <-
       sep = "\n",
       append = TRUE
     )
+    .write_main_package_R_file(DIR)
 
-    message("Updating DESCRIPTION ...", domain = NA)
-    .update_description_file(DIR)
+    message("Updating NAMESPACE ...", domain=NA)
+    suppressMessages(roxygen2::roxygenise(package.dir = DIR, clean = TRUE))
 
-    message("Updating Read-and-delete-me with Stan-specific notes...", domain = NA)
-    .update_read_and_delete_me(DIR)
+    message("Writing Read-and-delete-me file with additional instructions ...",
+            domain = NA)
+    .write_read_and_delete_me(DIR)
 
+    message("Finished skeleton for package: ", name, ".\n")
     message(
       domain = NA,
       sprintf(
@@ -222,31 +247,31 @@ rstan_package_skeleton <-
   file.path(base_url, relative_path)
 }
 
-.pkg_dependency <- function(pkg, last = FALSE) {
-  paste0(pkg, " (>= ", packageVersion(pkg), ")", if (!last) ", ")
-}
-
-.update_read_and_delete_me <- function(dir) {
+.write_read_and_delete_me <- function(dir) {
   cat(
-    "\n\nStan specific notes:\n",
-    "* Be sure to add useDynLib(mypackage, .registration = TRUE) to the NAMESPACE file, ",
-    "which you can do by placing the line   #' @useDynLib anRpackage, .registration = TRUE ",
-    "in one of your .R files (e.g., see rstanarm's 'rstanarm-package.R' file).",
-    "* Be sure to import all of Rcpp and methods in the NAMESPACE file.",
+    "* The precompiled stanmodel objects will appear in a named list called 'stanmodels', ",
+    "and you can call them with something like rstan::sampling(stanmodels$foo, ...)",
     "* You can put into src/stan_files/chunks any file that is needed by any .stan file in src/stan_files, ",
     "* You can put into inst/include any C++ files that are needed by any .stan file in src/stan_files, ",
     "but be sure to #include your C++ files in inst/include/meta_header.hpp",
-    "* The precompiled stanmodel objects will appear in a named list called 'stanmodels', ",
-    "and you can call them with something like rstan::sampling(stanmodels$foo, ...)",
+    "* While developing your package use devtools::install('.', local=FALSE) ",
+    "to reinstall the package AND recompile Stan programs, or set local=FALSE to skip the recompilation.",
     file = file.path(dir, "Read-and-delete-me"),
     sep = "\n",
-    append = TRUE
+    append = FALSE
   )
 }
 
 .update_description_file <- function(dir) {
+  available_pkgs <- available.packages(repos = "https://cran.rstudio.com/")[, c("Package", "Version")]
+  available_pkgs <- data.frame(available_pkgs, stringsAsFactors = FALSE)
+  .pkg_dependency <- function(pkg, last = FALSE) {
+    ver <- available_pkgs$Version[available_pkgs$Package == pkg]
+    paste0(pkg, " (>= ", ver, ")", if (!last) ", ")
+  }
+
   cat(
-    paste0("Depends: R (>= 3.3.0), ",
+    paste0("Depends: R (>= 3.4.0), ",
            .pkg_dependency("Rcpp"),
            "methods"),
     paste0("Imports: ",
@@ -258,8 +283,81 @@ rstan_package_skeleton <-
            .pkg_dependency("BH"),
            .pkg_dependency("Rcpp"),
            .pkg_dependency("RcppEigen", last=TRUE)),
+    "SystemRequirements: GNU make",
+    "NeedsCompilation: yes",
     file = file.path(dir, "DESCRIPTION"),
     sep = "\n",
     append = TRUE
+  )
+
+  DES <- readLines(file.path(dir, "DESCRIPTION"))
+  DES[grep("^License:", DES)] <- "License: GPL (>=3)"
+  cat(
+    DES,
+    file = file.path(dir, "DESCRIPTION"),
+    sep = "\n",
+    append = FALSE
+  )
+}
+
+.write_main_package_R_file <- function(dir) {
+  pkgname <- basename(dir)
+  cat(
+    paste0("#' The '", pkgname, "' package."),
+    "#' ",
+    "#' @description A DESCRIPTION OF THE PACKAGE",
+    "#' ",
+    "#' @docType package",
+    paste0("#' @name ", pkgname, "-package"),
+    paste0("#' @aliases ", pkgname),
+    paste0("#' @useDynLib ", pkgname, ", .registration = TRUE"),
+    "#' @import methods",
+    "#' @import Rcpp",
+    "#' @import rstantools",
+    "#' @importFrom rstan sampling",
+    "#' ",
+    "#' @references ",
+    paste0("#' ", .rstan_reference()),
+    "#' ",
+    "NULL",
+    file = file.path(dir, "R", paste0(pkgname, "-package.R")),
+    sep = "\n",
+    append = FALSE
+  )
+}
+
+.remove_unused_files <- function(dir) {
+  pkgname <- basename(dir)
+
+  if (file.exists(file.path(dir, "NAMESPACE"))) {
+    file.remove(file.path(dir, "NAMESPACE"))
+  }
+
+  DATA <- file.path(dir, "data")
+  if (file.exists(file.path(DATA, "delete_data.rds"))) {
+    file.remove(file.path(DATA, "delete_data.rds"))
+  }
+  if (!length(list.files(DATA))) {
+    file.remove(DATA)
+  }
+
+  MAN <- file.path(dir, "man")
+  if (file.exists(file.path(MAN, "delete_data.Rd"))) {
+    file.remove(file.path(MAN, "delete_data.Rd"))
+  }
+  if (file.exists(file.path(MAN, paste0(pkgname, "-package.Rd")))) {
+    file.remove(file.path(MAN, paste0(pkgname, "-package.Rd")))
+  }
+}
+
+
+.rstan_reference <- function() {
+  has_version <- utils::packageDescription("rstan", fields = "Version")
+  version_year <- substr(utils::packageDescription("rstan", fields = "Date"), 1, 4)
+  paste0(
+    "Stan Development Team (", version_year,"). ",
+    "RStan: the R interface to Stan. ",
+    "R package version ", has_version, ". ",
+    "http://mc-stan.org"
   )
 }